requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.5
django-ninja==0.20.0
idna==3.4
oauthlib==3.2.2
Pillow==9.4.0
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.14
